+++ modules/os2forms_digital_post/src/Plugin/WebformHandler/WebformHandlerSF1601.php
@@ -110,6 +110,7 @@ final class WebformHandlerSF1601 extends WebformHandlerBase {
       '#required' => TRUE,
       '#default_value' => $this->configuration[self::MEMO_MESSAGE][static::RECIPIENT_ELEMENT] ?? NULL,
       '#options' => $availableElements,
+      '#empty_option' => $this->t('- Select -'),
     ];
 
     $availableElements = $this->getAttachmentElements();
@@ -119,6 +120,7 @@ final class WebformHandlerSF1601 extends WebformHandlerBase {
       '#required' => TRUE,
       '#default_value' => $this->configuration[self::MEMO_MESSAGE][static::ATTACHMENT_ELEMENT] ?? NULL,
       '#options' => $availableElements,
+      '#empty_option' => $this->t('- Select -'),
     ];
 
     $form[self::MEMO_MESSAGE][self::SENDER_LABEL] = [