+++ CHANGELOG.md
@@ -15,6 +15,9 @@ before starting to add changes. Use example [placed in the end of the page](#exa
   signature validation text position) from the *OS2Forms Attachment* element
   onto the *Digital Signature* webform handler. Includes an automatic
   migration update hook for existing webforms.
+
+## [5.1.0] 2026-06-03
+
 - [PR-326](https://github.com/OS2Forms/os2forms/pull/326)
   Updating ckeditor -> ckeditor5.
 - [PR-322](https://github.com/OS2Forms/os2forms/pull/322)
@@ -31,11 +34,13 @@ before starting to add changes. Use example [placed in the end of the page](#exa
   Updated code analysis script.
 - [PR-306](https://github.com/OS2Forms/os2forms/pull/306)
   Made digital signature text placement configurable.
-- [#251](https://github.com/OS2Forms/os2forms/issues/251)
-  Webform encrypt uninstall problem fix
 - git actions check
 - [PR-289](https://github.com/OS2Forms/os2forms/pull/289)
   Added required "Zoom control position" to map element
+- [#246](https://github.com/OS2Forms/os2forms/issues/246)
+  Adding Datafordeler address lookup
+- [#251](https://github.com/OS2Forms/os2forms/issues/251)
+  Updated Webform encrypt uninstall problem fix
 
 ## [5.0.0] 2025-11-18
 
@@ -432,7 +437,8 @@ f/OS-115_dawa_address
 - Security in case of vulnerabilities.
 ```
 
-[Unreleased]: https://github.com/OS2Forms/os2forms/compare/5.0.0...HEAD
+[Unreleased]: https://github.com/OS2Forms/os2forms/compare/5.1.0...HEAD
+[5.1.0]: https://github.com/OS2Forms/os2forms/compare/5.0.0...5.1.0
 [5.0.0]: https://github.com/OS2Forms/os2forms/compare/4.1.0...5.0.0
 [4.1.0]: https://github.com/OS2Forms/os2forms/compare/4.0.0...4.1.0
 [4.0.0]: https://github.com/OS2Forms/os2forms/compare/3.22.2...4.0.0